Fixed the animation a bit
This code can at least run, but it has trouble when switching from one animation to another (it seems like a single "random" frame is shown). I assumed this was the "empty picture" array that was not properly initilized to zero, but that didn't seem to help. Regardless, this frame should not be shown at all, so some debugging still needs to be done.

Anyway, I won't debug this further at the moment as I'm not certain if this is the memory structure I want to use, so I'll look into the main branch again for a while.